feat(http): move back to http package
