+++ net/http/client.go
@@ -1,0 +1,11 @@
+package http
+
+import (
+	"context"
+)
+
+// Client for HTTP.
+type Client[Req any, Res any] interface {
+	// Call with request.
+	Call(ctx context.Context, req *Req) (*Res, error)
+}
