+++ net/http/rpc/handler.go
@@ -7,14 +7,9 @@ import (
 
 	"github.com/alexfalkowski/go-service/errors"
 	"github.com/alexfalkowski/go-service/marshaller"
+	nh "github.com/alexfalkowski/go-service/net/http"
 	"github.com/alexfalkowski/go-service/net/http/content"
 )
-
-// Handler for HTTP.
-type Handler[Req any, Res any] interface {
-	// Handle func for request/response.
-	Handle(ctx Context, req *Req) (*Res, error)
-}
 
 var (
 	mux *http.ServeMux
@@ -27,14 +22,14 @@ func Register(mu *http.ServeMux, ma *marshaller.Map) {
 }
 
 // Handler for HTTP.
-func Handle[Req any, Res any](path string, handler Handler[Req, Res]) {
+func Handle[Req any, Res any](path string, handler nh.Handler[Req, Res]) {
 	h := func(res http.ResponseWriter, req *http.Request) {
-		ctx := newContext(req.Context(), req, res)
+		ctx := nh.NewContext(req.Context(), req, res)
 		ct := content.NewFromRequest(req)
 
 		m, err := ct.Marshaller(mar)
 		if err != nil {
-			writeError(ctx, errors.Prefix("rpc marshaller", err))
+			nh.WriteError(ctx, errors.Prefix("rpc marshaller", err))
 
 			return
 		}
@@ -43,7 +38,7 @@ func Handle[Req any, Res any](path string, handler Handler[Req, Res]) {
 
 		body, err := io.ReadAll(req.Body)
 		if err != nil {
-			writeError(ctx, errors.Prefix("rpc read", err))
+			nh.WriteError(ctx, errors.Prefix("rpc read", err))
 
 			return
 		}
@@ -54,21 +49,21 @@ func Handle[Req any, Res any](path string, handler Handler[Req, Res]) {
 		ptr := &rq
 
 		if err := m.Unmarshal(body, ptr); err != nil {
-			writeError(ctx, errors.Prefix("rpc unmarshal", err))
+			nh.WriteError(ctx, errors.Prefix("rpc unmarshal", err))
 
 			return
 		}
 
 		rs, err := handler.Handle(ctx, ptr)
 		if err != nil {
-			writeError(ctx, errors.Prefix("rpc handle", err))
+			nh.WriteError(ctx, errors.Prefix("rpc handle", err))
 
 			return
 		}
 
 		d, err := m.Marshal(rs)
 		if err != nil {
-			writeError(ctx, errors.Prefix("rpc marshal", err))
+			nh.WriteError(ctx, errors.Prefix("rpc marshal", err))
 
 			return
 		}
@@ -77,8 +72,4 @@ func Handle[Req any, Res any](path string, handler Handler[Req, Res]) {
 	}
 
 	mux.HandleFunc("POST "+path, h)
-}
-
-func writeError(ctx Context, err error) {
-	http.Error(ctx.Response(), err.Error(), Code(err))
 }
