+++ net/http/status.go
@@ -1,4 +1,4 @@
-package rpc
+package http
 
 import (
 	"errors"
