+++ transport/http/http_test.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/alexfalkowski/go-service/limiter"
 	"github.com/alexfalkowski/go-service/meta"
+	nh "github.com/alexfalkowski/go-service/net/http"
 	"github.com/alexfalkowski/go-service/net/http/rpc"
 	"github.com/alexfalkowski/go-service/test"
 	v1 "github.com/alexfalkowski/go-service/test/greet/v1"
@@ -34,7 +35,7 @@ type Response struct {
 
 type SuccessHandler struct{}
 
-func (*SuccessHandler) Handle(ctx rpc.Context, r *Request) (*Response, error) {
+func (*SuccessHandler) Handle(ctx nh.Context, r *Request) (*Response, error) {
 	name := ctx.Request().URL.Query().Get("name")
 	if name == "" {
 		name = r.Name
@@ -47,18 +48,18 @@ func (*SuccessHandler) Handle(ctx rpc.Context, r *Request) (*Response, error) {
 
 type ProtobufHandler struct{}
 
-func (*ProtobufHandler) Handle(_ rpc.Context, r *v1.SayHelloRequest) (*v1.SayHelloResponse, error) {
+func (*ProtobufHandler) Handle(_ nh.Context, r *v1.SayHelloRequest) (*v1.SayHelloResponse, error) {
 	return &v1.SayHelloResponse{Message: "Hello " + r.GetName()}, nil
 }
 
-func (*ProtobufHandler) Error(_ rpc.Context, err error) *v1.SayHelloResponse {
+func (*ProtobufHandler) Error(_ nh.Context, err error) *v1.SayHelloResponse {
 	return &v1.SayHelloResponse{Message: err.Error()}
 }
 
 type ErrorHandler struct{}
 
-func (*ErrorHandler) Handle(_ rpc.Context, _ *Request) (*Response, error) {
-	return nil, rpc.Error(http.StatusServiceUnavailable, "ohh no")
+func (*ErrorHandler) Handle(_ nh.Context, _ *Request) (*Response, error) {
+	return nil, nh.Error(http.StatusServiceUnavailable, "ohh no")
 }
 
 func TestSync(t *testing.T) {
@@ -314,8 +315,8 @@ func TestDisallowedSync(t *testing.T) {
 				_, err := client.Call(context.Background(), &Request{Name: "Bob"})
 
 				Convey("Then I should have an error", func() {
-					So(rpc.IsError(err), ShouldBeTrue)
-					So(rpc.Code(err), ShouldEqual, http.StatusUnauthorized)
+					So(nh.IsError(err), ShouldBeTrue)
+					So(nh.Code(err), ShouldEqual, http.StatusUnauthorized)
 					So(err.Error(), ShouldContainSubstring, "verify token: invalid token")
 				})
